pyproject.toml
[tool.poetry]
name = "NMEA-Mux"
version = "0.1.0"
description = ""
authors = ["krgough <krgough@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.13.7"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
